some prestation check + when workedHour == plannedHour, the task is finished

+++ trunk/Timesheet/src/be/model/Prestation.java
@@ -50,7 +50,7 @@ public class Prestation extends AbstractTask {
 	 * retourne les heures prévues de la tâche
 	 */
 	@Override
-	public Integer getPlannedHours() {
+	public double getPlannedHours() {
 		return this.getParent().getPlannedHours();
 	}
